Moving As to separate namespace
