+++ src/With/With.csproj
@@ -25,4 +25,7 @@
     <PackageReleaseNotes>Adding net45 as target framework and fix for partition null reference exception bug</PackageReleaseNotes>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Coercions\" />
+  </ItemGroup>
 </Project>
