+++ src/With/WithExtension.cs
@@ -3,17 +3,32 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using With.WithPlumbing;
-using NameAndValue = System.Collections.Generic.KeyValuePair<string, object>;
 
 namespace With
 {
+    /// <summary>
+    /// With extensions. Used to create 
+    /// </summary>
     public static class WithExtensions
     {
+        /// <summary>
+        /// Starts constructing another object based on a base object.
+        /// </summary>
+        /// <returns>A builder for getting a new object based on the existing object.</returns>
+        /// <param name="t">The object to use as a template.</param>
+        /// <typeparam name="T">The type of the object that should be returned.</typeparam>
         public static ValuesForConstructor<T> With<T>(this T t)
         {
             return new ValuesForConstructor<T>(t);
         }
-
+        /// <summary>
+        /// Constructs another object based on a base object.
+        /// </summary>
+        /// <returns>A new object based on the existing object.</returns>
+        /// <param name="t">The object to use as a template.</param>
+        /// <param name="expr">What property or field to be different</param>
+        /// <param name="val">The value for the property expected to be changed</param>
+        /// <typeparam name="T">The type of the object that should be returned.</typeparam>
         public static T With<T, TVal>(this T t, Expression<Func<T, TVal>> expr, TVal val)
         {
             var memberAccess = new ExpressionWithMemberAccess();
@@ -21,12 +36,18 @@ namespace With
 
             return CreateInstanceFromValues.Create<T,T>(t,new[] { GetNameAndValue.Get(t, memberAccess.Members, val) });
         }
-
+        /// <summary>
+        /// Constructs another object based on a base object.
+        /// </summary>
+        /// <returns>A new object based on the existing object.</returns>
+        /// <param name="t">The object to use as a template.</param>
+        /// <param name="parameters">The constructor parameters to override.</param>
+        /// <typeparam name="T">The type of the object that should be returned.</typeparam>
         public static T With<T>(this T t, IDictionary<string, object> parameters)
         {
             return CreateInstanceFromValues.Create<T,T>(t,parameters);
         }
-
+/*
         public static T With<T>(this T t, Expression<Func<T, bool>> expr)
         {
             var eqeq = new ExpressionWithEqualEqualOrCall<T>(t);
@@ -50,30 +71,6 @@ namespace With
             var propertyNameAndValues = eqeq.Parsed.ToArray();
             return CreateInstanceFromValues.Create<T, T>(t,propertyNameAndValues);
         }
-
-        public static ValuesForConstructor<TRet> As<TRet>(this Object t)
-        {
-            return new ValuesForConstructor<TRet>(t);
-        }
-
-        public static TRet As<TRet>(this Object t, IDictionary<string, object> parameters)
-        {
-            return CreateInstanceFromValues.Create<TRet>(t, parameters.Select(v => new NameAndValue(v.Key, v.Value)));
-        }
-
-        public static TRet As<TRet>(this Object t, Expression<Func<TRet, object>> expr, object val)
-        {
-            var memberAccess = new ExpressionWithMemberAccess();
-            memberAccess.Lambda<TRet, object>(expr);
-            return CreateInstanceFromValues.Create<TRet>(t,new[] { GetNameAndValue.Get(t, memberAccess.Members, val) });
-        }
-
-        public static TRet As<TRet>(this Object t, Expression<Func<TRet, bool>> expr)
-        {
-            var eqeq = new ExpressionWithEqualEqualOrCall<TRet>(t);
-            eqeq.Lambda(expr);
-            var propertyNameAndValues = eqeq.Parsed.ToArray();
-            return CreateInstanceFromValues.Create<TRet>(t, propertyNameAndValues);
-        }
+*/
     }
 }
